// ui3/src/generated/bridge/SetAvatarBasePayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AvatarColor3 } from "./AvatarColor3";

export type SetAvatarBasePayload = { skinColor?: AvatarColor3, eyesColor?: AvatarColor3, hairColor?: AvatarColor3, bodyShapeUrn: string, name: string, };