// ui3/src/generated/bridge/SetAvatarPayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AvatarColor3 } from "./AvatarColor3";
import type { SetAvatarBasePayload } from "./SetAvatarBasePayload";
import type { SetAvatarEquipPayload } from "./SetAvatarEquipPayload";

export type SetAvatarPayload = { base?: SetAvatarBasePayload, equip?: SetAvatarEquipPayload, has_claimed_name?: boolean, profile_extras?: Record<string, unknown>, name_color?: AvatarColor3 | null, };